Revert "Update JOSDK to 5.1.0"
This reverts commit 9787195215845ae92cc0d54247ebfdd6636cf0ce.

Reason for revert: Update of existing installations causes issues with CronJob reconciliation

+++ operator/src/main/java/com/google/gerrit/k8s/operator/api/model/network/NetworkMember.java
@@ -19,7 +19,7 @@ import java.util.Objects;
 
 public class NetworkMember {
   private String name;
-  private long httpPort = 8080L;
+  private int httpPort = 8080;
 
   public NetworkMember() {}
 
@@ -36,11 +36,11 @@ public class NetworkMember {
     this.name = name;
   }
 
-  public long getHttpPort() {
+  public int getHttpPort() {
     return httpPort;
   }
 
-  public void setHttpPort(long httpPort) {
+  public void setHttpPort(int httpPort) {
     this.httpPort = httpPort;
   }
